# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GSI Partner Opportunity Tracker API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/org-export:
    get:
      operationId: orgExport
      tags: [orgData]
      summary: Export this org's complete dataset (admin only)
      responses:
        "200":
          description: Org export bundle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgExportBundle"
        "403":
          description: Admin access required
  /admin/org-import:
    post:
      operationId: orgImport
      tags: [orgData]
      summary: Import an org export bundle, additive (admin only)
      parameters:
        - name: dryRun
          in: query
          schema:
            type: string
            enum: ["true", "false"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgExportBundle"
      responses:
        "200":
          description: Import result or dry-run preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgImportResult"
        "400":
          description: Invalid bundle
        "403":
          description: Admin access required
...
components:
  schemas:
    OrgExportBundle:
      type: object
      required: [version, exportedAt, org, data]
      properties:
        version:
          type: string
        exportedAt:
          type: string
        org:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
        data:
          type: object
          additionalProperties:
            type: array
            items:
              type: object
              additionalProperties: true
    OrgImportResult:
      type: object
      properties:
        ok:
          type: boolean
        dryRun:
          type: boolean
        message:
          type: string
        counts:
          type: object
          additionalProperties:
            type: integer
        preview:
          type: object
          additionalProperties:
            type: integer
        warnings:
          type: array
          items:
            type: string
...